Added workaround for message queue event not triggering on Linux.
For some TBD reason, the message queue event is not triggering
which prevents poll from exiting in a timely fashion.  The
workaround is to reduce the poll timeout interval.  The
correct fix is to fix the underlying event not functioning
as expect.  The workaround will slightly increase latency
and processor overhead.