Fixed intermittent timeout broken for API calls.
The jsdrv api_cmd function uses a local stack variable for the
timeout structure.  It passes this to the frontend, which clears
it from the message.

The api_cmd was also later clearing the timeout structure from the pointer.
However, it was doing this after the message was passed to msg_queue_push,
which relinquishes ownership.

The problem occurs if an api_cmd invocation ended up getting
allocated the same message used by another active api_cmd call
from a different thread.  One of the threads could clear the timeout
structure pointer in the still active message, which converts
the command into asynchronous (no timeout).  This prevents the
command from ever completing, hanging the calling thread.